main_part1.py
# Fixed main.py - Part 1
import os
import cv2
import pickle
import base64
import logging
import numpy as np
import torch
import time
from datetime import datetime, timedelta
from typing import List, Optional, Any, Dict, Tuple
import httpx
import threading
import asyncio
import json
from concurrent.futures import ThreadPoolExecutor
from fastapi import FastAPI, File, UploadFile, HTTPException, Depends, WebSocket, WebSocketDisconnect, Form, Request, Response, status, Query
from fastapi.responses import HTMLResponse, RedirectResponse
from fastapi.templating import Jinja2Templates
from fastapi.staticfiles import StaticFiles
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from pydantic import BaseModel
from sqlalchemy import create_engine, Column, Integer, String, LargeBinary, DateTime, Boolean, Float, Text
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker, Session
from sqlalchemy.sql import func
import jwt
from passlib.context import CryptContext
from facenet_pytorch import MTCNN, InceptionResnetV1
from scipy.spatial import distance as dist
from scipy.optimize import linear_sum_assignment
from collections import OrderedDict, defaultdict
from filterpy.kalman import KalmanFilter
import dlib
from imutils import face_utils
import skimage.transform
from sqlalchemy import inspect

# ----------------------------------------------------------------------------------
# DATABASE SETUP
# ----------------------------------------------------------------------------------
DB_FOLDER = "db"
if not os.path.exists(DB_FOLDER):
    os.makedirs(DB_FOLDER)

SQLALCHEMY_DATABASE_URL = f"sqlite:///{DB_FOLDER}/app.db"
engine = create_engine(SQLALCHEMY_DATABASE_URL, connect_args={"check_same_thread": False})
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
Base = declarative_base()

TEMPLATES_DIR = "templates"
if not os.path.exists(TEMPLATES_DIR):
    os.makedirs(TEMPLATES_DIR)

STATIC_DIR = "static"
if not os.path.exists(STATIC_DIR):
    os.makedirs(STATIC_DIR)

LOG_DIR = "/app/logs"
if not os.path.exists(LOG_DIR):
    os.makedirs(LOG_DIR)
    try:
        os.chmod(LOG_DIR, 0o777)
    except PermissionError:
        pass

try:
    with open('/tmp/logs/app.log', 'a') as f:
        pass
    log_file = '/tmp/logs/app.log'
except (PermissionError, FileNotFoundError):
    log_file = None

logging.basicConfig(
    level=logging.INFO,
    format="[%(asctime)s] %(levelname)s: %(message)s",
    handlers=[
        logging.StreamHandler()
    ] + ([logging.FileHandler(log_file)] if log_file else [])
)
logger = logging.getLogger("FaceRecognitionAPI")

templates = Jinja2Templates(directory=TEMPLATES_DIR)

# ----------------------------------------------------------------------------------
# DATABASE MODELS
# ----------------------------------------------------------------------------------
class User(Base):
    __tablename__ = "users"
    id = Column(Integer, primary_key=True, index=True)
    username = Column(String, unique=True, index=True)
    hashed_password = Column(String)

class ConfigEntry(Base):
    __tablename__ = "config"
    key = Column(String, primary_key=True, index=True)
    value = Column(String)

class RegisteredFace(Base):
    __tablename__ = "registered_faces"
    id = Column(Integer, primary_key=True, index=True)
    person_id = Column(String, unique=True, index=True)
    embedding = Column(LargeBinary)
    openface_embedding = Column(LargeBinary)
    hog_features = Column(LargeBinary)
    updated_at = Column(DateTime, default=datetime.utcnow)
    confidence = Column(Float)
    face_quality = Column(Float)
    landmarks = Column(LargeBinary)

class TrackingHistory(Base):
    __tablename__ = "tracking_history"
    id = Column(Integer, primary_key=True, index=True)
    person_id = Column(String, index=True)
    track_id = Column(Integer, index=True)
    timestamp = Column(DateTime, default=datetime.utcnow)
    position_x = Column(Float)
    position_y = Column(Float)
    velocity_x = Column(Float)
    velocity_y = Column(Float)
    confidence = Column(Float)
    bbox_x = Column(Float)
    bbox_y = Column(Float)
    bbox_w = Column(Float)
    bbox_h = Column(Float)

class PersonTrajectory(Base):
    __tablename__ = "person_trajectories"
    id = Column(Integer, primary_key=True, index=True)
    person_id = Column(String, index=True)
    track_id = Column(Integer, index=True)
    start_time = Column(DateTime)
    end_time = Column(DateTime)
    trajectory_data = Column(Text)
    total_distance = Column(Float)
    avg_speed = Column(Float)

class RTSPStream(Base):
    __tablename__ = "rtsp_streams"
    id = Column(Integer, primary_key=True, index=True)
    name = Column(String, index=True)
    rtsp_url = Column(String)
    is_active = Column(Boolean, default=False)
    created_at = Column(DateTime, default=datetime.utcnow)
    last_detection = Column(DateTime)
    detection_count = Column(Integer, default=0)
    avg_processing_time = Column(Float)

class DetectionEvent(Base):
    __tablename__ = "detection_events"
    id = Column(Integer, primary_key=True, index=True)
    person_id = Column(String, index=True)
    stream_id = Column(Integer, index=True)
    timestamp = Column(DateTime, default=datetime.utcnow)
    confidence = Column(Float)
    bbox_x = Column(Float)
    bbox_y = Column(Float)
    bbox_w = Column(Float)
    bbox_h = Column(Float)
    face_quality = Column(Float)

# Temporary placeholder
if __name__ == "__main__":
    pass
