requirements.txt
dlib
fastapi==0.95.1
uvicorn==0.22.0
gunicorn==20.1.0
sqlalchemy==1.4.47
PyJWT==2.6.0
passlib[bcrypt]==1.7.4
opencv-python==4.6.0.66
numpy==1.24.2
torch==1.13.1
facenet-pytorch==2.5.2
python-multipart==0.0.5
Jinja2==3.1.2
httpx
uvicorn[standard]
openface
filterpy
scipy
imutils
scikit-image